# src/docs/swagger.yaml
basePath: /api/v2
definitions:
  consts.Language:
    enum:
    - am
    - ar
    - as
    - af
    - bn
    - zh
    - da
    - nl
    - en
    - fi
    - fr
    - de
    - el
    - gu
    - ha
    - hi
    - ig
    - it
    - ja
    - kn
    - ko
    - ml
    - mr
    - ne
    - "no"
    - or
    - om
    - fa
    - pt
    - pa
    - ru
    - sa
    - st
    - sn
    - sd
    - so
    - es
    - sw
    - sv
    - ta
    - te
    - ts
    - tn
    - tr
    - ur
    - '-'
    - ve
    - xh
    - yo
    - zu
    type: string
    x-enum-varnames:
    - Amharic
    - Arabic
    - Assamese
    - Afrikaans
    - Bangla
    - Chinese
    - Danish
    - Dutch
    - English
    - Finnish
    - French
    - German
    - Greek
    - Gujarati
    - Hausa
    - Hindi
    - Igbo
    - Italian
    - Japanese
    - Kannada
    - Korean
    - Malayalam
    - Marathi
    - Nepali
    - Norwegian
    - Odia
    - Oromo
    - Persian
    - Portuguese
    - Punjabi
    - Russian
    - Sanskrit
    - Sesotho
    - Shona
    - Sindhi
    - Somali
    - Spanish
    - Swahili
    - Swedish
    - Tamil
    - Telugu
    - Tsonga
    - Tswana
    - Turkish
    - Urdu
    - Undefined
    - Venda
    - Xhosa
    - Yoruba
    - Zulu
  database.Campaign:
    properties:
      campaignId:
        type: string
      createdAt:
        description: read only
        type: string
      createdById:
        type: string
      description:
        type: string
      endDate:
        type: string
      image:
        type: string
      language:
        $ref: '#/definitions/consts.Language'
      name:
        type: string
      rules:
        type: string
      startDate:
        type: string
    type: object
  database.CampaignRound:
    properties:
      allowCreations:
        type: boolean
      allowExpansions:
        type: boolean
      allowJuryToParticipate:
        type: boolean
      allowMultipleJudgement:
        type: boolean
      allowedMediaTypes:
        items:
          $ref: '#/definitions/database.MediaType'
        type: array
      blacklist:
        type: string
      campaignId:
        type: string
      createdAt:
        type: string
      createdById:
        type: string
      dependsOnRoundId:
        type: string
      description:
        type: string
      endDate:
        type: string
      isOpen:
        type: boolean
      isPublic:
        type: boolean
      maximumSubmissionOfSameArticle:
        type: integer
      minimumAddedBytes:
        type: integer
      minimumAddedWords:
        type: integer
      minimumDurationMilliseconds:
        type: integer
      minimumHeight:
        type: integer
      minimumResolution:
        type: integer
      minimumTotalBytes:
        type: integer
      minimumTotalWords:
        type: integer
      minimumWidth:
        type: integer
      name:
        type: string
      roundId:
        type: string
      secretBallot:
        type: boolean
      serial:
        type: integer
      startDate:
        type: string
      totalSubmissions:
        type: integer
    type: object
  database.MediaType:
    enum:
    - ARTICLE
    - BITMAP
    - AUDIO
    - VIDEO
    - PDF
    type: string
    x-enum-varnames:
    - MediaTypeArticle
    - MediaTypeImage
    - MediaTypeAudio
    - MediaTypeVideo
    - MediaTypePDF
  routes.ResponseList-database_Campaign:
    properties:
      data:
        items:
          $ref: '#/definitions/database.Campaign'
        type: array
    type: object
  routes.ResponseList-database_CampaignRound:
    properties:
      data:
        items:
          $ref: '#/definitions/database.CampaignRound'
        type: array
    type: object
  routes.ResponseSingle-database_CampaignRound:
    properties:
      data:
        $ref: '#/definitions/database.CampaignRound'
    type: object
  routes.ResponseSingle-services_RoundImportSummary:
    properties:
      data:
        $ref: '#/definitions/services.RoundImportSummary'
    type: object
  services.CampaignCreateRequest:
    properties:
      allowCreations:
        type: boolean
      allowExpansions:
        type: boolean
      allowJuryToParticipate:
        type: boolean
      allowMultipleJudgement:
        type: boolean
      allowedMediaTypes:
        items:
          $ref: '#/definitions/database.MediaType'
        type: array
      blacklist:
        type: string
      description:
        type: string
      endDate:
        type: string
      image:
        type: string
      jury:
        items:
          type: string
        type: array
      language:
        $ref: '#/definitions/consts.Language'
      maximumSubmissionOfSameArticle:
        type: integer
      minimumAddedBytes:
        type: integer
      minimumAddedWords:
        type: integer
      minimumDurationMilliseconds:
        type: integer
      minimumHeight:
        type: integer
      minimumResolution:
        type: integer
      minimumTotalBytes:
        type: integer
      minimumTotalWords:
        type: integer
      minimumWidth:
        type: integer
      name:
        type: string
      rules:
        type: string
      secretBallot:
        type: boolean
      startDate:
        type: string
    type: object
  services.CampaignUpdateRequest:
    properties:
      allowCreations:
        type: boolean
      allowExpansions:
        type: boolean
      allowJuryToParticipate:
        type: boolean
      allowMultipleJudgement:
        type: boolean
      allowedMediaTypes:
        items:
          $ref: '#/definitions/database.MediaType'
        type: array
      blacklist:
        type: string
      description:
        type: string
      endDate:
        type: string
      image:
        type: string
      jury:
        items:
          type: string
        type: array
      language:
        $ref: '#/definitions/consts.Language'
      maximumSubmissionOfSameArticle:
        type: integer
      minimumAddedBytes:
        type: integer
      minimumAddedWords:
        type: integer
      minimumDurationMilliseconds:
        type: integer
      minimumHeight:
        type: integer
      minimumResolution:
        type: integer
      minimumTotalBytes:
        type: integer
      minimumTotalWords:
        type: integer
      minimumWidth:
        type: integer
      name:
        type: string
      rules:
        type: string
      secretBallot:
        type: boolean
      startDate:
        type: string
    type: object
  services.ImportFromCommonsPayload:
    properties:
      categories:
        description: Categories from which images will be fetched
        items:
          type: string
        type: array
    type: object
  services.ImportStatus:
    enum:
    - success
    - failed
    - pending
    type: string
    x-enum-varnames:
    - ImportStatusSuccess
    - ImportStatusFailed
    - ImportStatusPending
  services.RoundImportSummary:
    properties:
      failedCount:
        type: integer
      failedIds:
        items:
          type: string
        type: array
      status:
        $ref: '#/definitions/services.ImportStatus'
      successCount:
        type: integer
    type: object
  services.RoundRequest:
    properties:
      allowCreations:
        type: boolean
      allowExpansions:
        type: boolean
      allowJuryToParticipate:
        type: boolean
      allowMultipleJudgement:
        type: boolean
      allowedMediaTypes:
        items:
          $ref: '#/definitions/database.MediaType'
        type: array
      blacklist:
        type: string
      campaignId:
        type: string
      dependsOnRoundId:
        type: string
      description:
        type: string
      endDate:
        type: string
      isOpen:
        type: boolean
      isPublic:
        type: boolean
      maximumSubmissionOfSameArticle:
        type: integer
      minimumAddedBytes:
        type: integer
      minimumAddedWords:
        type: integer
      minimumDurationMilliseconds:
        type: integer
      minimumHeight:
        type: integer
      minimumResolution:
        type: integer
      minimumTotalBytes:
        type: integer
      minimumTotalWords:
        type: integer
      minimumWidth:
        type: integer
      name:
        type: string
      secretBallot:
        type: boolean
      serial:
        type: integer
      startDate:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: nokibsarkar@gmail.com
    name: Nokib Sarkar
    url: https://github.com/nokibsarkar
  description: This is the API documentation for Campwiz
  license:
    name: GPL-3.0
    url: https://www.gnu.org/licenses/gpl-3.0.html
  title: Campwiz API
  version: "1"
paths:
  /campaign/:
    get:
      description: get all campaigns
      parameters:
      - collectionFormat: csv
        in: query
        items:
          type: string
        name: ids
        type: array
      - in: query
        name: limit
        type: integer
      - in: query
        name: nextToken
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/routes.ResponseList-database_Campaign'
      summary: List all campaigns
      tags:
      - Campaign
    post:
      description: Create a new campaign
      parameters:
      - description: The campaign request
        in: body
        name: campaignRequest
        required: true
        schema:
          $ref: '#/definitions/services.CampaignCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/database.Campaign'
      summary: Create a new campaign
      tags:
      - Campaign
  /campaign/{id}:
    post:
      description: Update a campaign
      parameters:
      - description: The campaign ID
        in: path
        name: id
        required: true
        type: string
      - description: The campaign request
        in: body
        name: campaignRequest
        required: true
        schema:
          $ref: '#/definitions/services.CampaignUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/database.Campaign'
      summary: Update a campaign
      tags:
      - Campaign
  /round/:
    get:
      description: get all rounds
      parameters:
      - in: query
        name: campaignId
        type: string
      - in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/routes.ResponseList-database_CampaignRound'
      summary: List all rounds
      tags:
      - Round
    post:
      description: Create a new round for a campaign
      parameters:
      - description: The round request
        in: body
        name: roundRequest
        required: true
        schema:
          $ref: '#/definitions/services.RoundRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/routes.ResponseSingle-database_CampaignRound'
      summary: Create a new round
      tags:
      - Round
  /round/import/{roundId}:
    get:
      description: It would be used as a server sent event stream to broadcast on
        the frontend about current status of the round
      parameters:
      - description: The round ID
        in: path
        name: roundId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/routes.ResponseSingle-services_RoundImportSummary'
      summary: Get the import status about a round
      tags:
      - Round
    post:
      description: The user would provide a round ID and a list of commons categories
        and the system would import images from those categories
      parameters:
      - description: The round ID
        in: path
        name: roundId
        required: true
        type: string
      - description: The import from commons request
        in: body
        name: ImportFromCommons
        required: true
        schema:
          $ref: '#/definitions/services.ImportFromCommonsPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/routes.ResponseSingle-services_RoundImportSummary'
      summary: Import images from commons
      tags:
      - Round
produces:
- application/json
schemes:
- http
- https
swagger: "2.0"